+++ lib/data_frame.ex
@@ -47,6 +47,7 @@ defmodule DataFrame do
   @doc """
     Creates a Frame from the textual output of a frame (allows copying data from webpages, etc.)
   """
+  @spec parse(String.t) :: Frame.t
   def parse(text) do
     [header | data ] = String.split(text, "\n", trim: true)
     columns = String.split(header, " ", trim: true)
@@ -74,19 +75,51 @@ defmodule DataFrame do
   end
 
   # ##################################################
-  #  Ordering
+  #  Transforming and Sorting
   # ##################################################
 
   @doc """
     Returns a Frame which data has been transposed.
   """
+  @spec transpose(Frame.t) :: Frame.t
   def transpose(frame) do
     %Frame{values: Table.transpose(frame.values), index: frame.columns, columns: frame.index}
+  end
+
+
+  @doc """
+  Creates a list of Dataframes grouped by one of the columns.
+  A , B
+  1 , 2
+  1,  3
+  2, 4
+  group_by(A)
+  [ A B
+    1 2
+    1 3,
+    A B
+    2 4
+  ]
+  """
+  def group_by(frame, master_column) do
+    frame
+    |> column(master_column)
+    |> Enum.uniq
+    |> Enum.map(fn(value) -> filter_rows(frame, master_column, value) end)
+  end
+
+  @doc """
+  DataFrame.to_list_of_maps DataFrame.new([[1,2],[3,4]], ["A", "B"])
+  > [%{"A" => 1, "B" => 2}, %{"A" => 3, "B" => 4}]
+  """
+  def to_list_of_maps(frame) do
+    # TODO
   end
 
   @doc """
     Sorts the data in the frame based on its index. By default the data is sorted in ascending order.
   """
+  @spec sort_index(Frame.t, boolean) :: Frame.t
   def sort_index(frame, ascending \\ true) do
     sort(frame, 0, ascending)
   end
@@ -94,6 +127,7 @@ defmodule DataFrame do
   @doc """
     Sorts the data in the frame based on a given column. By default the data is sorted in ascending order.
   """
+  @spec sort_values(Frame.t, String.t, boolean) :: Frame.t
   def sort_values(frame, column_name, ascending \\ true) do
     index = Enum.find_index(frame.columns, fn(x) -> x == column_name end)
     sort(frame, index + 1, ascending)
@@ -120,6 +154,7 @@ defmodule DataFrame do
   @doc """
     Returns the information at the top of the frame. Defaults to 5 lines.
   """
+  @spec head(Frame.t, integer) :: Frame.t
   def head(frame, size \\ 5) do
     DataFrame.new(Enum.take(frame.values, size), frame.columns, Enum.take(frame.index, size))
   end
@@ -127,11 +162,37 @@ defmodule DataFrame do
   @doc """
     Returns the information at the bottom of the frame. Defaults to 5 lines.
   """
+  @spec tail(Frame.t, integer) :: Frame.t
   def tail(frame, the_size \\ 5) do
     size = -the_size
     head(frame, size)
   end
 
+  @doc """
+  Returns a Frame with the selected columns by name. similar to loc
+  # TODO: merge with loc?
+  """
+  def columns(frame, column_names) do
+    column_index = frame.columns |> Enum.with_index |> Enum.map( fn(tuple) ->
+      if Enum.find_index(column_names, fn(name) -> name == elem(tuple, 0) end) != nil do
+        elem(tuple, 1)
+      end
+    end)
+
+    column_indexes = Enum.filter(column_index, fn(x) -> x != nil end)
+
+    columns = at_a(frame.columns, column_indexes)
+    values = Table.columns_index(frame.values, column_indexes)
+    DataFrame.new(values, columns, frame.index)
+  end
+
+  # TODO: move somewhere
+  # same than .at but accepting a list of indexes
+  defp at_a(list, list_index) do
+    Enum.map(list_index, fn(index) -> Enum.at(list, index) end)
+  end
+
+  @spec column(Frame.t, String.t) :: Frame.t
   def column(frame, column_name) do
     column = Enum.find_index(frame.columns, fn(x) -> to_string(x) == to_string(column_name) end)
     frame.values |> Table.transpose |> Enum.at(column)
@@ -141,6 +202,7 @@ defmodule DataFrame do
     Returns a slice of the data in the frame.
     Parameters are the ranges with names in the index and column
   """
+  #@spec loc(Frame.t, Range.t, Range.t) :: Frame.t
   def loc(frame, index_range, column_range) do
     DataFrame.iloc(frame, index_range_integer(frame, index_range), column_range_integer(frame, column_range))
   end
@@ -177,6 +239,7 @@ defmodule DataFrame do
   @doc """
     Returns a value located at the position indicated by an index name and column name.
   """
+  @spec at(Frame.t, String.t, String.t) :: any()
   def at(frame, index_name, column_name) do
     index = Enum.find_index(frame.index, fn(x) -> to_string(x) == to_string(index_name) end)
     column = Enum.find_index(frame.columns, fn(x) -> to_string(x) == to_string(column_name) end)
@@ -186,8 +249,78 @@ defmodule DataFrame do
   @doc """
     Returns a value located at the position indicated by an index position and column position.
   """
+  @spec iat(Frame.t, integer, integer) :: any()
   def iat(frame, index, column) do
-    Table.at(frame.values, column, index)
+    Table.at(frame.values, index, column)
+  end
+
+
+  @doc """
+    Experimental
+    Returns the rows that contains certain value in a column
+    # TODO: rationalize all this slicing operations
+  """
+  def filter_rows(frame, expected_column_name, expected_value) do
+    column_index = Enum.find_index(frame.columns, fn(x) -> x == expected_column_name end)
+    if column_index == nil do
+      frame
+    else
+      values = Enum.map(frame.values,
+       fn(row) ->
+         if Enum.at(row, column_index) == expected_value do
+           row
+         else
+           [nil]
+         end
+       end
+     )
+     {new_values, new_index} = delete_nil_rows(values, frame.index)
+     DataFrame.new(new_values, frame.columns, new_index)
+    end
+
+  end
+
+  @doc """
+  Experimental
+  Returns a frame with the info for which `fun` returned true. Extremely greedy. Only elements, not rows/columns
+  """
+  def filter(frame, fun) do
+    with_nils = Enum.map(Table.with_index(frame.values), fn(row_tuple) ->
+      row = elem(row_tuple, 0)
+      row_index = elem(row_tuple, 1)
+      row_name = Enum.at(frame.index, row_index)
+      Enum.map row, fn(column_tuple) ->
+        value = elem(column_tuple, 0)
+        column_index = elem(column_tuple, 1)
+        column_name = Enum.at(frame.columns, column_index)
+        if fun.(value, column_name, column_index, row_name, row_index) do
+          value
+        else
+          nil
+        end
+      end
+    end)
+    {new_table, new_index} = delete_nil_rows(with_nils, frame.index)
+    new_columns = frame.columns
+    {final_table, new_columns} = delete_nil_rows(Table.transpose(new_table), frame.columns)
+    result_table = if final_table == [[]] do
+        [[]]
+      else
+         Table.transpose(final_table)
+        end
+    DataFrame.new(new_table, new_columns, new_index)
+  end
+
+  defp delete_nil_rows([], _) do
+    {[[]], []}
+  end
+  defp delete_nil_rows(table, list) do
+    nil_index = Enum.find_index(table, fn(row) -> Enum.all?(row, fn(element) -> element == nil end) end)
+    if nil_index == nil do
+      {table, list}
+    else
+      delete_nil_rows(List.delete_at(table, nil_index), List.delete_at(list, nil_index))
+    end
   end
 
   # ##################################################
@@ -197,6 +330,7 @@ defmodule DataFrame do
   @doc """
     Returns the cummulative sum
   """
+  @spec cumsum(Frame.t) :: Frame.t
   def cumsum(frame) do
     columns = frame.values |> Table.transpose
     cumsummed = columns |> Enum.map( fn(column) ->
@@ -211,6 +345,7 @@ defmodule DataFrame do
   @doc """
     Returns a statistical description of the data in the frame
   """
+  @spec describe(Frame.t) :: Frame.t
   def describe(frame) do
     DataFrame.Statistics.describe(frame)
   end
@@ -235,11 +370,13 @@ defmodule DataFrame do
   @doc """
     Reads the information from a CSV file. By default the first row is assumed to be the column names.
   """
+  @spec from_csv(String.t) :: Frame.t
   def from_csv(filename) do
     [headers | values] = filename |> File.stream! |> CSV.decode |> Enum.to_list
     new(values, headers)
   end
 
+  @spec plot(Frame.t) :: :ok
   def plot(frame) do
     plotter = Explot.new
     columns_with_index = frame.values |> Table.transpose |> Enum.with_index